+++ src/xch/providers/leaflet_rpc/leaflet_rpc_provider.ts
@@ -2,15 +2,18 @@ import { BigNumber } from "@ethersproject/bignumber";
 import { Network } from "../../../util/network";
 import { SpendBundle } from "../../../util/serializer/types/spend_bundle";
 import { Provider, getBalanceArgs, subscribeToPuzzleHashUpdatesArgs, subscribeToCoinUpdatesArgs, getPuzzleSolutionArgs, getCoinChildrenArgs, getBlockHeaderArgs, getBlocksHeadersArgs, getCoinRemovalsArgs, getCoinAdditionsArgs } from "../provider";
-import { pushSpendBundleArgs, transferArgs, transferCATArgs, acceptOfferArgs, subscribeToAddressChangesArgs, signCoinSpendsArgs, changeNetworkArgs } from "../provider_args";
-import { Optional, PuzzleSolution, CoinState, BlockHeader, Coin } from "../provider_types";
+import { pushSpendBundleArgs, transferArgs, transferCATArgs, acceptOfferArgs, subscribeToAddressChangesArgs, signCoinSpendsArgs, changeNetworkArgs, getCoinsArgs } from "../provider_args";
+import { Optional, PuzzleSolution, CoinState, BlockHeader, Coin, bytes } from "../provider_types";
 import axios from "axios";
+import { Util } from "../../../util";
+import { uint } from "../../../util/serializer/basic_types";
 
 export class LeafletRPCProvider implements Provider {
     public baseUrl: string;
 
     private network: Network;
     private connected = false;
+    private overwriteGetRPCResponse: (<T>(endpoint: string, params: any) => Promise<T | null>) | null;
 
     constructor(
         host: string,
@@ -18,9 +21,11 @@ export class LeafletRPCProvider implements Provider {
         port = 18444,
         network = Network.mainnet,
         useHTTP = false,
+        overwriteGetRPCResponse: (<T>(endpoint: string, params: any) => Promise<T | null>) | null = null
     ) {
         this.baseUrl = (useHTTP ? "http://" : "https://") + `${host}:${port}/${apiKey}/rpc/`;
         this.network = network;
+        this.overwriteGetRPCResponse = overwriteGetRPCResponse;
     }
 
     public async connect(): Promise<void> {
@@ -61,14 +66,92 @@ export class LeafletRPCProvider implements Provider {
                 peak: {
                     height: number
                 }
-            }
+            },
+            success: boolean
         }>("get_blockchain_state", {});
 
+        if(!resp?.success) return null;
         return resp?.blockchain_state.peak.height ?? null;
     }
 
-    getBalance(args: getBalanceArgs): Promise<Optional<BigNumber>> {
-        throw new Error("Method not implemented.");
+    public async getBalance({ address, puzzleHash, minHeight = 1 }: getBalanceArgs): Promise<Optional<BigNumber>> {
+        const cs: CoinState[] | null = await this.getCoins({
+            address,
+            puzzleHash,
+            startHeight:
+            minHeight,
+            includeSpentCoins: false
+        });
+        if(cs === null) return null;
+        
+        let balance: BigNumber = BigNumber.from(0);
+
+        for(const coinState of cs) {
+            balance = balance.add(coinState.coin.amount);
+        }
+
+        return balance;
+    }
+
+    public async getCoins({
+        address,
+        puzzleHash,
+        startHeight,
+        endHeight,
+        includeSpentCoins
+    }: getCoinsArgs): Promise<Optional<CoinState[]>> {
+        let puzHash: string;
+        
+        // get puzHash: Buffer from address / puzzle hash
+        if(address !== undefined) {
+            puzHash = Util.address.addressToPuzzleHash(address);
+            if(puzHash.length === 0) {
+                return null;
+            }
+        }
+        else if(puzzleHash !== undefined) {
+            puzHash = Util.address.validateHashString(puzzleHash);
+        }
+        else return null;
+
+        const reqParams: any = { puzzle_hash: puzHash };
+        if(startHeight !== undefined) { reqParams.start_height = startHeight }
+        if(endHeight !== undefined) { reqParams.end_height = endHeight }
+        if(includeSpentCoins !== undefined) { reqParams.include_spent_coins = includeSpentCoins; }
+
+        const resp = await this.getRPCResponse<{
+            success: boolean,
+            coin_records: Array<{
+                coin: {
+                    amount: uint,
+                    parent_coin_info: bytes,
+                    puzzle_hash: bytes
+                },
+                confirmed_block_index: uint | null,
+                spent_block_index: uint | null
+            }>,
+        }>("get_coin_records_by_puzzle_hash", reqParams);
+
+        if(!resp?.success) return null;
+        const cs: CoinState[] = [];
+
+        for(const cr of resp.coin_records) {
+            const c = new Coin();
+            c.amount = BigNumber.from(cr.coin.amount);
+            // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+            c.parentCoinInfo = Util.dehexlify(cr.coin.parent_coin_info)!;
+            // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+            c.puzzleHash = Util.dehexlify(cr.coin.puzzle_hash)!;
+
+            const coinState = new CoinState();
+            coinState.coin = c;
+            coinState.createdHeight = cr.confirmed_block_index;
+            coinState.spentHeight = cr.spent_block_index;
+
+            cs.push(coinState);
+        }
+
+        return cs;
     }
 
     subscribeToPuzzleHashUpdates(args: subscribeToPuzzleHashUpdatesArgs): void {
@@ -107,31 +190,36 @@ export class LeafletRPCProvider implements Provider {
         throw new Error("Method not implemented.");
     }
 
-    getAddress(): Promise<string> {
-        throw new Error("Method not implemented.");
+    private _doesNotImplementError(): any {
+        throw new Error("LeafletRPCProvider does not implement this method.");
     }
 
-    transfer(args: transferArgs): Promise<Optional<SpendBundle>> {
-        throw new Error("Method not implemented.");
+    public getAddress(): Promise<string> {
+        return this._doesNotImplementError();
     }
 
-    transferCAT(args: transferCATArgs): Promise<Optional<SpendBundle>> {
-        throw new Error("Method not implemented.");
+    public transfer(args: transferArgs): Promise<Optional<SpendBundle>> {
+        return this._doesNotImplementError();
     }
 
-    acceptOffer(args: acceptOfferArgs): Promise<Optional<SpendBundle>> {
-        throw new Error("Method not implemented.");
+    public transferCAT(args: transferCATArgs): Promise<Optional<SpendBundle>> {
+        return this._doesNotImplementError();
     }
 
-    subscribeToAddressChanges(args: subscribeToAddressChangesArgs): void {
-        throw new Error("Method not implemented.");
+    public acceptOffer(args: acceptOfferArgs): Promise<Optional<SpendBundle>> {
+        return this._doesNotImplementError();
     }
 
-    signCoinSpends(args: signCoinSpendsArgs): Promise<Optional<SpendBundle>> {
-        throw new Error("Method not implemented.");
+    public subscribeToAddressChanges(args: subscribeToAddressChangesArgs): void {
+        return this._doesNotImplementError();
+    }
+
+    public signCoinSpends(args: signCoinSpendsArgs): Promise<Optional<SpendBundle>> {
+        return this._doesNotImplementError();
     }
 
     public async changeNetwork({ network }: changeNetworkArgs): Promise<boolean> {
         this.network = network;
+        return true;
     }
 }