+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - make `Singleton` available to apps as `greenweb.Singleton`
  - rename old `LeafletProvider` to `LeafletWSProvider`
  - create `LeafletRPCProvider` and ensure that it's the default `LeafletProvider` for `greenweb.xch.createProvider` (to overwrite, set the `useLeafletWS` parameter to `true`)
+ - `Provider`: add the `getCoins()` method
  - `greenweb.sexp`:
      - add new helper method: `bytesToAtom()`
      - add the `singletonPuzzle()` and `singletonSolution()` methods
@@ -24,7 +25,7 @@
       - add the `useP2SingletonCoinsConditionsAndCoinSol()` method
       - add tests
  - add more tests to ensure full code coverage (100%)
- - todo: update docs to reflect changes
+ - update docs to reflect changes
 
 ## 1.1.8
  - `PrivateKeyProvider`: the `signCoinSpends` now accepts a second, optional argument: `customGenesisChallenge`